+++ forward_engineering/helpers/buildScript/joinScripts.js
@@ -1,0 +1,7 @@
+const joinScripts = structure => {
+	return structure.filter(item => item).join('\n\n');
+};
+
+module.exports = {
+	joinScripts,
+};
